// src/utils/enums.ts
export enum Tile {
  EARTH = "earth",
  WATER = "water",
}

export enum Water {
  RIVER = "river",
  LAKE = "lake",
  SEA = "sea",
}

export enum WaterType {
  SHALLOW = "shallow",
  MEDIUM = "medium",
  DEEP = "deep",
}

export enum EarthType {
  FIELD = "field",
  SAND = "sand",
  FOREST = "forest",
  MOUNTAIN = "montain",
}